pyproject.toml
[project]
name = "sail"
version = "0.1.0"
description = "A well formatted, easy-to-use implementation of https://arxiv.org/pdf/2412.17092"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.1",
    "openai>=1.58.1",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "scikit-learn>=1.6.0",
    "sentence-transformers>=3.3.1",
    "tiktoken>=0.8.0",
]
